create custom hook: useSeedBankSummary
